// src/models/cards.model.ts
export enum cardsLabels {
    channelLabel = 'channelLabel',
    showConnectionInfoButtonLabel = 'showConnectionInfoButtonLabel',
    download = 'download',
    upload = 'upload',
    providerHeaderLabel = 'providerHeaderLabel',
    moreDetailsLabel = 'moreDetailsLabel',
    hiddenDetailsLabel = 'hiddenDetailsLabel',
    applicationsCardTitleLabel = 'applicationsCardTitleLabel',
    systemCardTitleLabel = 'systemCardTitleLabel',
    homeWiFiCardTitleLabel = 'homeWiFiCardTitleLabel',
    guestWiFiCardTitleLabel = 'guestWiFiCardTitleLabel',
    networkPortsCardTitleLabel = 'networkPortsCardTitleLabel',
    internetCardTitleLabel = 'internetCardTitleLabel',
    hasNoApplicationsMessageLabel = 'hasNoApplicationsMessageLabel',
    onlineDeviceTitleLabel = 'onlineDeviceTitleLabel',
    wifiConnectionTypeMessageLabel = 'wifiConnectionTypeMessageLabel',
    wiredConnectionTypeMessageLabel = 'wiredConnectionTypeMessageLabel',
    connectedLabel = 'connectedLabel',
    daysLabel = 'daysLabel',

    modelLabel = 'modelLabel',
    serviceTagLabel = 'serviceTagLabel',
    osVersionLabel = 'osVersionLabel',
    autoUpdateLabel = 'autoUpdateLabel',
    upTimeLabel = 'upTimeLabel',
    currentTimeLabel = 'currentTimeLabel',
    cpuUsageLabel = 'cpuUsageLabel',
    ramUsageLabel = 'ramUsageLabel',

    enabledLabel = 'enabledLabel',
    kbitsLabel = 'kbitsLabel',
    mbitsLabel = 'mbitsLabel',
    mbitLabel = 'mbitLabel',

    usingEthernetLabel = 'usingEthernetLabel',

    statusLabel = 'statusLabel',
    internetSafetyLabel = 'internetSafetyLabel',
    authenticationTypeLabel = 'authenticationTypeLabel',
    ipAddressLabel = 'ipAddressLabel',
    subnetMaskLabel = 'subnetMaskLabel',
    macAddressLabel = 'macAddressLabel',
    downloadLabel = 'downloadLabel',
    uploadLabel = 'uploadLabel',
    receivedLabel = 'receivedLabel',
    sentLabel = 'sentLabel',
    dnsServersLabel = 'dnsServersLabel',

    connectionEstablishedLabel = 'connectionEstablishedLabel',
    rebootLabel = 'rebootLabel',
    configureLabel = 'configureLabel',
    adGuardIsEnabledLabel = 'adGuardIsEnabledLabel',
}